# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Skykin platform API — developer portal (SDK keys), ad campaign portal (advertisers/operators), SDK event ingestion, intent prediction, campaign ad delivery via WebSocket, and reward notifications.",
        "title": "Skykin Platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/ad-portal/admin/users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ad Portal - Admin"
                ],
                "summary": "Create portal user (operator admin)",
                "parameters": [
                    {
                        "description": "User",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_advertisers_interfaces_http.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_advertisers_interfaces_http.RegisterResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        },
...
        "/ad-portal/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ad Portal - Auth"
                ],
                "summary": "Login to ad portal",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_advertisers_interfaces_http.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_advertisers_interfaces_http.AdPortalLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        },
        "/ad-portal/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ad Portal - Auth"
                ],
                "summary": "Current advertiser profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_advertisers_interfaces_http.MeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        },
        "/ad-portal/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ad Portal - Auth"
                ],
                "summary": "Register advertiser",
                "parameters": [
                    {
                        "description": "Registration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_advertisers_interfaces_http.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_advertisers_interfaces_http.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        },
...
        "/portal/applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all applications belonging to the authenticated developer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal - Applications"
                ],
                "summary": "List applications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a new application under the authenticated developer and generates SDK API key credentials. The secret key is only shown once.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal - Applications"
                ],
                "summary": "Create a new application",
                "parameters": [
                    {
                        "description": "Application details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_auth_dto.ApplicationCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        },
        "/portal/login": {
            "post": {
                "description": "Authenticates a developer and returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal - Auth"
                ],
                "summary": "Login developer",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_auth_dto.DeveloperLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        },
        "/portal/register": {
            "post": {
                "description": "Creates a new developer account for the Skykin portal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal - Auth"
                ],
                "summary": "Register a new developer",
                "parameters": [
                    {
                        "description": "Developer registration details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_auth_dto.DeveloperRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_advertisers_interfaces_http.AdPortalLoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/internal_advertisers_interfaces_http.PortalUserDTO"
                }
            }
        },
        "internal_advertisers_interfaces_http.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "role"
            ],
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "advertiser",
                        "read_only_analyst",
                        "operator_admin"
                    ]
                }
            }
        },
        "internal_advertisers_interfaces_http.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_advertisers_interfaces_http.MeResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/internal_advertisers_interfaces_http.PortalUserDTO"
                }
            }
        },
        "internal_advertisers_interfaces_http.PortalUserDTO": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string",
                    "example": "abc123..."
                },
                "company_name": {
                    "type": "string",
                    "example": "Acme Inc"
                },
                "contact_name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "email": {
                    "type": "string",
                    "example": "advertiser@test.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "role": {
                    "type": "string",
                    "example": "advertiser"
                }
            }
        },
        "internal_advertisers_interfaces_http.RegisterRequest": {
            "type": "object",
            "required": [
                "company_name",
                "email",
                "name",
                "password"
            ],
            "properties": {
                "company_name": {
                    "type": "string",
                    "example": "Acme Inc"
                },
                "email": {
                    "type": "string",
                    "example": "advertiser@test.com"
                },
                "name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "SecurePass1!"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "advertiser",
                        "read_only_analyst"
                    ],
                    "example": "advertiser"
                }
            }
        },
        "internal_advertisers_interfaces_http.RegisterResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/internal_advertisers_interfaces_http.PortalUserDTO"
                }
            }
        },
...
        "skykin-platform_internal_auth_dto.ApplicationCreateRequest": {
            "type": "object",
            "required": [
                "app_name",
                "bundle_id",
                "platform"
            ],
            "properties": {
                "app_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "My Shopping App"
                },
                "bundle_id": {
                    "type": "string",
                    "example": "com.company.app"
                },
                "platform": {
                    "type": "string",
                    "enum": [
                        "flutter",
                        "android",
                        "ios",
                        "web"
                    ],
                    "example": "flutter"
                }
            }
        },
        "skykin-platform_internal_auth_dto.DeveloperLoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jane@company.com"
                },
                "password": {
                    "type": "string",
                    "example": "securepass123"
                }
            }
        },
        "skykin-platform_internal_auth_dto.DeveloperRegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jane@company.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Jane Developer"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "securepass123"
                }
            }
        },
...
        "skykin-platform_internal_platform_http.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "skykin-platform_internal_platform_http.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "SDK publishable key (pk_live_...)",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Enter your JWT token as: Bearer \u003ctoken\u003e (developer portal or ad portal)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}